Clean Header and Notifications
