+++ package.json
@@ -14,7 +14,6 @@
     "esbuild": "^0.25.2",
     "primeicons": "^7.0.0",
     "primereact": "^10.9.3",
-    "prop-types": "^15.8.1",
     "pusher-js": "^8.4.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
